Add `_logger` field to WorldCommand class (incomplete)
A new private field `_logger` of type `Logger` was added to the
`WorldCommand` class in `WorldCommand.cs`. However, the initialization
of `_logger` is incomplete, as the assignment is left unfinished
(`Logger.`). This change appears to be a work in progress or requires
further implementation to properly initialize the logger instance.
---
Ensafe HitPayload.cs to prevent a buff bug
Checked the entire Act I and annotated
Created some helper methods
Code cleanup